Add files via upload
We need to guess a number between 1 to 100. The computer will ask you to guess a number.
You will get some chances to guess the number. If you will guess the number in the given number of times. You will get the message that "Won".
You can add more Guesses so a person can make more guesses. In this case, you can make 5 guesses.

It is a small interesting game to play.